+++ crates/anneal-lang/src/loader.rs
@@ -303,8 +303,7 @@ fn collect_statement_definitions(statement: &Statement, definitions: &mut BTreeS
         | Statement::Import(_)
         | Statement::Verb(_)
         | Statement::Doc(_)
-        | Statement::Predicate(_)
-        | Statement::Cookbook(_) => {}
+        | Statement::Predicate(_) => {}
     }
 }
 
@@ -344,8 +343,7 @@ fn qualify_statement(
         | Statement::Include(_)
         | Statement::Import(_)
         | Statement::Verb(_)
-        | Statement::Doc(_)
-        | Statement::Cookbook(_) => {}
+        | Statement::Doc(_) => {}
     }
 }